generate 2 abstract classes for bed
