+++ pygenometracks/tests/test_data/master_tracks.ini
@@ -132,21 +132,11 @@ height = 2
 # Otherwise, each plot use its own scale
 #overlay_previous = yes
 
-# If the bed file contains the exon
-# structure (bed 12) then this is plotted. Otherwise
-# a region **with direction** is plotted.
-# If the bed file contains a column for color (column 9), then this color can be used by
-# setting:
-#color = bed_rgb
-# if color is a valid colormap name (like RbBlGn), then the score (column 5) is mapped
-# to the colormap.
-# In this case, the the min_value and max_value for the score can be provided, otherwise
-# the maximum score and minimum score found are used.
-#color = RdYlBu
-#min_value=0
-#max_value=100
-# If the color is simply a color name, then this color is used and the score is not considered.
-color = darkblue
+# optional: line_width
+#line_width = 0.5
+# optional, default is black. To remove the border, simply set 'border_color' to none
+#border_color = black
+
 # whether printing the labels
 labels = false
 # optional:
@@ -155,8 +145,6 @@ labels = false
 #max_labels = 60
 # optional: font size can be given to override the default size
 fontsize = 10
-# optional: line_width
-#line_width = 0.5
 # the display parameter defines how the bed file is plotted.
 # Default is 'stacked' where regions are plotted on different lines so
 # we can see all regions and all labels.
@@ -164,9 +152,6 @@ fontsize = 10
 # These options assume that the regions do not overlap.
 # `collapsed`: The bed regions are plotted one after the other in one line.
 # `interleaved`: The bed regions are plotted in two lines, first up, then down, then up etc.
-# optional, default is black. To remove the border, simply set 'border_color' to none
-# Not used in tssarrow style
-#border_color = black
 # style to plot the genes when the display is not triangles
 #style = UCSC
 #style = flybase
@@ -205,9 +190,22 @@ fontsize = 10
 # If you want that the tip of the arrow correspond to
 # the extremity of the interval use:
 # arrowhead_included = true
+
+# If the bed file contains a column for color (column 9), then this color can be used by
+# setting:
+#color = bed_rgb
+# if color is a valid colormap name (like RbBlGn), then the score (column 5) is mapped
+# to the colormap.
+# In this case, the the min_value and max_value for the score can be provided, otherwise
+# the maximum score and minimum score found are used.
+#color = RdYlBu
+#min_value=0
+#max_value=100
+# If the color is simply a color name, then this color is used and the score is not considered.
+color = darkblue
 # optional. If not given is guessed from the file ending.
 file_type = bed
-    
+
 [epilog.qcat]
 file = pygenometracks/tests/test_data/epilog.qcat.bgz
